Same for isdigit() being int->int
git-svn-id: https://dev.csoft.net/agar/trunk@9444 86a84d31-9702-0410-8f1e-a60cab7bd424

+++ rg/tileset.c
@@ -969,7 +969,7 @@ tryname2:
 					    (np-ins_tile_name)-1, "%d", num);
 					break;
 				}
-				if (!isdigit(*np)) {
+				if (!isdigit((int) *np)) {
 					Strlcat(ins_tile_name, "_",
 					    sizeof(ins_tile_name));
 					break;
@@ -1076,7 +1076,7 @@ tryname2:
 					    (np-ins_texture_name)-1, "%d", num);
 					break;
 				}
-				if (!isdigit(*np)) {
+				if (!isdigit((int) *np)) {
 					Strlcat(ins_texture_name, "_",
 					    sizeof(ins_texture_name));
 					break;
@@ -1143,7 +1143,7 @@ tryname2:
 					    (np-ins_anim_name)-1, "%d", num);
 					break;
 				}
-				if (!isdigit(*np)) {
+				if (!isdigit((int) *np)) {
 					Strlcat(ins_anim_name, "_",
 					    sizeof(ins_anim_name));
 					break;
